# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * InsureWise API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HandoffRequest {
  userProfileId: number;
  policyId: string;
  planName: string;
  monthlyPremium: number;
}

export type HandoffResultStatus =
  (typeof HandoffResultStatus)[keyof typeof HandoffResultStatus];

export const HandoffResultStatus = {
  success: "success",
  fallback: "fallback",
} as const;

export interface HandoffResult {
  status: HandoffResultStatus;
  checkoutUrl: string;
  expiresIn?: number;
  message?: string;
}

export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface PriorityWeights {
  /**
   * @minimum 0
   * @maximum 100
   */
  price: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  coverage: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  rating: number;
}

export interface VehicleDetails {
  make: string;
  model: string;
  year: number;
  vin?: string;
}

export type PropertyDetailsType =
  (typeof PropertyDetailsType)[keyof typeof PropertyDetailsType];

export const PropertyDetailsType = {
  house: "house",
  condo: "condo",
  apartment: "apartment",
  townhouse: "townhouse",
} as const;

export interface PropertyDetails {
  address: string;
  type: PropertyDetailsType;
  yearBuilt?: number;
  squareFeet?: number;
}

export type UserProfileInsuranceType =
  (typeof UserProfileInsuranceType)[keyof typeof UserProfileInsuranceType];

export const UserProfileInsuranceType = {
  auto: "auto",
  home: "home",
  life: "life",
  health: "health",
  renters: "renters",
} as const;

export interface UserProfile {
  id: string;
  name: string;
  age: number;
  location: string;
  budgetMonthly: number;
  insuranceType: UserProfileInsuranceType;
  priorities: PriorityWeights;
  requirements: string[];
  vehicleDetails?: VehicleDetails;
  propertyDetails?: PropertyDetails;
  onboardingComplete: boolean;
  createdAt: string;
  updatedAt: string;
}

export type UpsertUserProfileRequestInsuranceType =
  (typeof UpsertUserProfileRequestInsuranceType)[keyof typeof UpsertUserProfileRequestInsuranceType];

export const UpsertUserProfileRequestInsuranceType = {
  auto: "auto",
  home: "home",
  life: "life",
  health: "health",
  renters: "renters",
} as const;

export interface UpsertUserProfileRequest {
  name: string;
  age: number;
  location: string;
  budgetMonthly: number;
  insuranceType: UpsertUserProfileRequestInsuranceType;
  priorities: PriorityWeights;
  requirements: string[];
  vehicleDetails?: VehicleDetails;
  propertyDetails?: PropertyDetails;
  onboardingComplete?: boolean;
}

export type CoverageItemStatus =
  (typeof CoverageItemStatus)[keyof typeof CoverageItemStatus];

export const CoverageItemStatus = {
  covered: "covered",
  partial: "partial",
  not_covered: "not_covered",
} as const;

export interface CoverageItem {
  type: string;
  name: string;
  status: CoverageItemStatus;
  details: string;
  limit?: string;
}

export interface PolicyCard {
  id: string;
  insurerName: string;
  insurerLogo?: string;
  planName: string;
  monthlyPremium: number;
  annualPremium: number;
  deductible: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  matchScore: number;
  priceScore: number;
  coverageScore: number;
  ratingScore: number;
  /**
   * @minimum 0
   * @maximum 5
   */
  overallRating: number;
  reviewCount: number;
  coverageSummary: CoverageItem[];
  gapCount: number;
  highlights: string[];
  warnings: string[];
}

export type SearchPoliciesRequestInsuranceType =
  (typeof SearchPoliciesRequestInsuranceType)[keyof typeof SearchPoliciesRequestInsuranceType];

export const SearchPoliciesRequestInsuranceType = {
  auto: "auto",
  home: "home",
  life: "life",
  health: "health",
  renters: "renters",
} as const;

export interface SearchPoliciesRequest {
  userProfileId?: string;
  insuranceType: SearchPoliciesRequestInsuranceType;
  priorities: PriorityWeights;
  requirements: string[];
  budgetMonthly: number;
  location: string;
}

export interface SearchPoliciesResponse {
  policies: PolicyCard[];
  totalFound: number;
  searchDuration: number;
}

export type CoverageExplanationStatus =
  (typeof CoverageExplanationStatus)[keyof typeof CoverageExplanationStatus];

export const CoverageExplanationStatus = {
  covered: "covered",
  partial: "partial",
  not_covered: "not_covered",
} as const;

export interface CoverageExplanation {
  requirement: string;
  status: CoverageExplanationStatus;
  explanation: string;
  policyClause?: string;
  /**
   * @minimum 0
   * @maximum 1
   */
  confidence: number;
}

export interface ExplainPolicyRequest {
  requirements: string[];
  userContext?: string;
}

export type ExplainPolicyResponseKeyTermsItem = {
  term: string;
  definition: string;
};

export interface ExplainPolicyResponse {
  policyId: string;
  summary: string;
  coverageItems: CoverageExplanation[];
  covered: string[];
  partial: string[];
  gaps: string[];
  recommendation: string;
  keyTerms: ExplainPolicyResponseKeyTermsItem[];
}

export interface GetApplicationFormRequest {
  userProfileId: string;
}

export type ApplicationFieldFieldType =
  (typeof ApplicationFieldFieldType)[keyof typeof ApplicationFieldFieldType];

export const ApplicationFieldFieldType = {
  text: "text",
  number: "number",
  date: "date",
  select: "select",
  boolean: "boolean",
} as const;

export interface ApplicationField {
  fieldId: string;
  label: string;
  value: string;
  fieldType: ApplicationFieldFieldType;
  options?: string[];
  required: boolean;
  editable: boolean;
}

export type ApplicationFormSectionsItem = {
  title: string;
  fields: ApplicationField[];
};

export interface ApplicationForm {
  policyId: string;
  insurerName: string;
  planName: string;
  monthlyPremium: number;
  sections: ApplicationFormSectionsItem[];
  autoFilledCount: number;
  totalFieldCount: number;
}

export type SubmitApplicationRequestFieldsItem = {
  fieldId: string;
  value: string;
};

export interface SubmitApplicationRequest {
  policyId: string;
  userProfileId: string;
  fields: SubmitApplicationRequestFieldsItem[];
}

export type ApplicationConfirmationStatus =
  (typeof ApplicationConfirmationStatus)[keyof typeof ApplicationConfirmationStatus];

export const ApplicationConfirmationStatus = {
  submitted: "submitted",
  pending_review: "pending_review",
  approved: "approved",
} as const;

export interface ApplicationConfirmation {
  confirmationId: string;
  policyNumber: string;
  insurerName: string;
  planName: string;
  startDate: string;
  monthlyPremium: number;
  coverageSummary: string[];
  status: ApplicationConfirmationStatus;
  message: string;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface AiChatRequest {
  messages: ChatMessage[];
  context?: string;
  extractProfile?: boolean;
}

export interface AiChatResponse {
  message: string;
  extractedProfile?: UpsertUserProfileRequest;
  isComplete: boolean;
  nextQuestion?: string;
}

export interface AiParseAnswerRequest {
  questionId: string;
  questionText: string;
  answer: string;
}

export type AiParseAnswerResponseExtractedEntities = { [key: string]: unknown };

export interface AiParseAnswerResponse {
  parsedValue: string;
  extractedEntities?: AiParseAnswerResponseExtractedEntities;
}