# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: InsureWise API specification
servers:
  - url: /api
    description: Base API path
...
paths:
  /checkout/handoff:
    post:
      operationId: initiateCarrierHandoff
      tags: [checkout]
      summary: Initiate automated carrier form handoff
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HandoffRequest"
      responses:
        "200":
          description: Carrier handoff result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HandoffResult"
...
components:
  schemas:
    HandoffRequest:
      type: object
      properties:
        userProfileId:
          type: integer
        policyId:
          type: string
        planName:
          type: string
        monthlyPremium:
          type: number
      required:
        - userProfileId
        - policyId
        - planName
        - monthlyPremium
    HandoffResult:
      type: object
      properties:
        status:
          type: string
          enum: [success, fallback]
        checkoutUrl:
          type: string
        expiresIn:
          type: integer
        message:
          type: string
      required:
        - status
        - checkoutUrl
...