# lib/api-zod/src/generated/types/handoffRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * InsureWise API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HandoffRequest {
  userProfileId: number;
  policyId: string;
  planName: string;
  monthlyPremium: number;
}