# lib/api-zod/src/generated/types/handoffResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * InsureWise API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HandoffResultStatus } from "./handoffResultStatus";

export interface HandoffResult {
  status: HandoffResultStatus;
  checkoutUrl: string;
  expiresIn?: number;
  message?: string;
}